+++ packages/cmsattribute/api/changesets.ts
@@ -1,0 +1,3 @@
+import type { PartialAttributeChangesets } from '@global/types/changesets';
+
+export const changesets: PartialAttributeChangesets = [];