+++ packages/cmsprevnext/src/init.ts
@@ -1,7 +1,7 @@
 import type { CMSItem, CMSList } from '@finsweet/attributes-cmscore';
 import { importCMSCore } from '@finsweet/attributes-cmscore';
 import { removeTrailingSlash } from '@finsweet/ts-utils';
-import { CMS_PREV_NEXT_ATTRIBUTE } from '@global/constants/attributes';
+import { CMS_ATTRIBUTE_ATTRIBUTE, CMS_PREV_NEXT_ATTRIBUTE } from '@global/constants/attributes';
 
 import { collectElements } from './collect';
 import { getSelector } from './constants';
@@ -12,6 +12,8 @@ import { getSelector } from './constants';
 export const init = async (): Promise<CMSList[]> => {
   const cmsCore = await importCMSCore();
   if (!cmsCore) return [];
+
+  await window.fsAttributes[CMS_ATTRIBUTE_ATTRIBUTE]?.loading;
 
   let previousPlaceholderFilled = false;
   let nextPlaceholderFilled = false;