+++ packages/cmssort/src/init.ts
@@ -1,6 +1,6 @@
 import type { CMSList } from '@finsweet/attributes-cmscore';
 import { importCMSCore } from '@finsweet/attributes-cmscore';
-import { CMS_SORT_ATTRIBUTE } from '@global/constants/attributes';
+import { CMS_ATTRIBUTE_ATTRIBUTE, CMS_SORT_ATTRIBUTE } from '@global/constants/attributes';
 
 import { initListSorting } from './factory';
 import { getSelector } from './utils/constants';
@@ -11,6 +11,8 @@ import { getSelector } from './utils/constants';
 export const init = async (): Promise<CMSList[]> => {
   const cmsCore = await importCMSCore();
   if (!cmsCore) return [];
+
+  await window.fsAttributes[CMS_ATTRIBUTE_ATTRIBUTE]?.loading;
 
   const listInstances = cmsCore.createCMSListInstances([getSelector('element', 'list', { operator: 'prefixed' })]);