+++ packages/cmstabs/src/init.ts
@@ -1,7 +1,7 @@
 import type { CMSList } from '@finsweet/attributes-cmscore';
 import { importCMSCore } from '@finsweet/attributes-cmscore';
 import { restartWebflow } from '@finsweet/ts-utils';
-import { CMS_TABS_ATTRIBUTE } from '@global/constants/attributes';
+import { CMS_ATTRIBUTE_ATTRIBUTE, CMS_TABS_ATTRIBUTE } from '@global/constants/attributes';
 
 import { collectPopulateData } from './actions/collect';
 import { populateTabsFromLists } from './actions/populate';
@@ -13,6 +13,8 @@ import { getSelector } from './utils/constants';
 export const init = async (): Promise<CMSList[]> => {
   const cmsCore = await importCMSCore();
   if (!cmsCore) return [];
+
+  await window.fsAttributes[CMS_ATTRIBUTE_ATTRIBUTE]?.loading;
 
   const listInstances = cmsCore.createCMSListInstances([getSelector('element', 'list', { operator: 'prefixed' })]);