+++ packages/rangeslider/src/init.ts
@@ -1,5 +1,5 @@
 import { getHiddenParent, isNotEmpty, isVisible } from '@finsweet/ts-utils';
-import { RANGE_SLIDER_ATTRIBUTE } from '@global/constants/attributes';
+import { CMS_ATTRIBUTE_ATTRIBUTE, RANGE_SLIDER_ATTRIBUTE } from '@global/constants/attributes';
 import debounce from 'just-debounce';
 
 import { getClientX } from './actions/events';
@@ -13,7 +13,9 @@ import type { HandleInstances } from './utils/types';
 /**
  * Inits the attribute.
  */
-export const init = (): HandleInstances[] => {
+export const init = async (): Promise<HandleInstances[]> => {
+  await window.fsAttributes[CMS_ATTRIBUTE_ATTRIBUTE]?.loading;
+
   const wrapperElements = [
     ...document.querySelectorAll<HTMLElement>(getSelector('element', 'wrapper', { operator: 'prefixed' })),
   ];