+++ global/import/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @global/import
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+
 ## 1.0.2
 
 ### Patch Changes