+++ packages/autovideo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @finsweet/attributes-autovideo
 
+## 1.4.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.3.4
 
 ### Patch Changes
@@ -59,11 +71,11 @@ It can be used like:
 <script>
   window.fsAttributes = window.fsAttributes || [];
   window.fsAttributes.push([
-    "autovideo",
-    videosState => {
-      console.log("Attribute has successfully loaded!");
+    'autovideo',
+    (videosState) => {
+      console.log('Attribute has successfully loaded!');
       console.log(videosState);
-    }
+    },
   ]);
 </script>
 ```