+++ packages/codehighlight/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @finsweet/attributes-codehighlight
 
+## 1.4.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.3.5
 
 ### Patch Changes