+++ packages/mirrorinput/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @finsweet/attributes-mirrorinput
 
+## 1.3.5
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.3.4
 
 ### Patch Changes
@@ -54,10 +62,10 @@ It can be used like:
 <script>
   window.fsAttributes = window.fsAttributes || [];
   window.fsAttributes.push([
-    "mirrorinput",
+    'mirrorinput',
     () => {
-      console.log("Attribute has successfully loaded!");
-    }
+      console.log('Attribute has successfully loaded!');
+    },
   ]);
 </script>
 ```