+++ packages/rangeslider/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @finsweet/attributes-rangeslider
 
+## 1.7.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.6.5
 
 ### Patch Changes
@@ -53,11 +65,11 @@ It can be used like:
 <script>
   window.fsAttributes = window.fsAttributes || [];
   window.fsAttributes.push([
-    "rangeslider",
-    handleInstances => {
-      console.log("Attribute has successfully loaded!");
+    'rangeslider',
+    (handleInstances) => {
+      console.log('Attribute has successfully loaded!');
       console.log(handleInstances);
-    }
+    },
   ]);
 </script>
 ```